Draft test on Repository classes
